Add C++ library to get full message definition for ros1 types. (#21)
**Public-Facing Changes**
None

**Description**
For ROS1, the full ros message definition of a topic is only available
when subscribing to the topic. In order to avoid that, a simple library
was added that allows to generate the full message schema without having
to subscribe to the topic. The library works similar to `roslib`'s
`gendeps` tool in that it parses `.msg` file on the disk to generate a
schema where all types are concatenated.

This library will be used to get the message schemas for published
topics. These will then be advertised as available channels.