Add ROS1 subscription support (#29)
**Public-Facing Changes**
Advertises ROS1 topics as channels and allows to subscribe to them e.g.
from Foxglove Studio.

**Description**
Implementation is mainly based on
https://github.com/achim-k/foxglove_websocket_server

One major problem that I noticed, is that the message schema parsing
approach (parsing .msg files from the disk) does not work if topics are
published from a bag file. We will have to think about what to do here,
maybe the solution is then to fallback and subscribe to the topic once
to get the message definition from the connection header.