Fix deadlock when unsubscribing from topic. (#49)
**Public-Facing Changes**
Fixes a deadlock that could occur when unsubscribing from a ROS1 topic

**Description**
The unsubscribe handler was called with exclusive lock on
`_clientsChannelMutex`. As a consequence, incoming messages on that
topic were waiting on this mutex and the `ros::Subscriber` object could
not be destroyed as the `ros::SubscriptionQueue's` mutex was locked.
With this patch, the unsubscribe handler is called without holding a
exclusive lock on `_clientsChannelMutex`.

I didn't test if the ros2 version was also affected, but I would guess
so.