[ROS2] Remove max_update_ms parameter, poll for rosgraph changes (#52)
**Public-Facing Changes**
- `max_update_ms` parameter is removed from the ROS 2 bridge, graph
changes are detected immediately

**Description**
Instead of a refresh loop with a timer, we run a thread that polls for
rosgraph changes using `wait_for_graph_change()`. The addition or
removal of topics is detected immediately instead of waiting up to five
seconds.

This approach only works for ROS 2. In ROS 1, the `roscore` service will
only broadcast publisher changes for existing topics to subscribers of
that topic. New topics can only be discovered by calling an XML-RPC
method, one of `getPublishedTopics`, `getTopicTypes`, or
`getSystemState`.

Fixes #15.