Add support for latched topics (or topics published with transient local policy) (#53)
**Public-Facing Changes**
Add support for topics that are latched (ROS1) or published with
transient local durability policy (ROS2)

**Description**
Instead of having only a single subscriber instance per topic, we now
create a subscriber for every client that subscribes to the topic. This
way we do not have to re-implement the latching logic and we instead
rely on ROS to correctly send latched messages to new subscriber
instances.

Fixes #30 
Fixes #31