Add SSL server support (#57)
**Public-Facing Changes**
Allows to start the server with SSL enabled


**Description**
This PR adds TLS support which can be enabled by setting the
corresponding `tls` parameter.

Excerpt from README:

> * __tls__: If `true`, use Transport Layer Security (TLS) for encrypted
communication. Defaults to `false`.
> * __certfile__: Path to the certificate to use for TLS. Required when
__tls__ is set to `true`. Defaults to `""`.
> * __keyfile__: Path to the private key to use for TLS. Required when
__tls__ is set to `true`. Defaults to `""`.

I could not get this working with a self signed certificate. Only made
it work with a certificate generated with let's encrypt (added an entry
in `/etc/hosts` to resolve the name to my local docker container)

Fixes #14