Refactor ROS 2 node as a component (#63)
**Public-Facing Changes**
Refactor ROS 2 node as a component


**Description**
Allows to run the ROS2 version both as standalone node or as a
component.

For the standalone version we simply create a `ComponentManager` and
load the foxglove bridge component. I'm not sure if this is best
practice, but it has the advantage that it does not additional header
files to be included which would require an additional refactoring.


Fixes #28