Keep one callback group per channel in memory (#92)
**Public-Facing Changes**
None

**Description**
Previously, we created a callback group when a client subscribed to a
topic and deleted the callback group again when the client unsubscribed.
This caused some side effects however (see #84) as ROS2 apparently does
not handle destruction of callback groups correctly. This PR ensures
that callback groups stay alive for the lifetime of the node.
Furthermore, we have one callback group per channel rather than one
callback group for every subscription