Remove "/usr/lib/cmake/" hardcoded path in CMakeLists.txt (#96)
**Public-Facing Changes**
- None

**Description**
Fixes CMake warnings on Ubuntu Bionic by removing the platform-specific
search path for nlohmann_json's CMake file, allowing it to fail on
Bionic quietly, and building anyways. This will result in a compile-time
error instead of a configuration-time error for users who don't have the
correct system dependency installed, but the error is easy enough to
deduce what went wrong and we print a status message warning this may
happen.