Drop messages when connection send buffer limit has been reached (#126)
**Public-Facing Changes**
- Drop messages when connection send buffer limit has been reached


**Description**
We currently do have the problem that slow clients/connections can cause
the server's connection send buffer to grow indefinitely potentially
leading to the server being killed as it consumes too much memory. This
PR introduces a user configurable send buffer limit. If the limit has
been reached, messages that are to be send over this connection will be
dropped in order to prevent the send buffer growing indefinitely.

See also #116