Always --pull when building docker images in the makefile (#140)
**Public-Facing Changes**
None

**Description**
Previously, running a test command through the Makefile such as `make
rolling-test` would use a previously cached ROS docker image if
available. This means the tests would
execute with an arbitrarily old version of the ROS distro. This adds
`--pull` to the build commands so new ROS base images are always pulled
if available before building
and running tests.