Run client handler functions in separate thread (#165)
**Public-Facing Changes**
- Run client handlers in separate thread


**Description**
Builds on top of #156 

So far we have executed client handler functions in the websocket server
thread. This could cause issues with clients not receiving messages
anymore while a longer-running handler function was executing. This PR
fixes this by handler functions being executed in a separate thread.

Fixes #164 
Fixes FG-2029