Fix publishers being created with invalid QoS profile (#205)
### Public-Facing Changes
- Fix publishers being created with invalid QoS profile

### Description
Since #184 we use the QoS profile of an existing publisher when creating
a new publisher for a client advertised topic. In case the existing
publisher has been destroyed before (but is still registered, see #204),
the QoS profile is invalid and creating a publisher fails with the
message `'create_publisher() failed setting topic QoS, at
./src/publisher.cpp:226'`.

This PR mitigates the issue by taking the QoS profile of an existing
publisher from a different node (if there is any). The assumption is,
that publishers of other nodes are less likely to be destroyed.