Extract ROS 2 bridge header (#228)
### Public-Facing Changes

Extracts ROS 2 bridge header such that the node class can be used
directly, not only as a component.

### Description

I want to compile and use the bridge in
https://github.com/mvukov/rules_ros2. I don't have support for
components, yet, so I'd like to use the bridge library itself. For that
reason, I extracted the header file.

There are no functional changes.