Fix QoS history being unknown when copied from existing publisher (#233)
### Public-Facing Changes
Fix QoS history being unknown when copied from existing publisher

### Description
When the QoS profile is copied from another existing publisher, it can
happen that the history policy is `Unknown`, leading to an error when
subsequently trying to create a publisher with that QoS profile. As a
fix, we explicitly set the history policy to the system default.

Fixes https://github.com/foxglove/ros-foxglove-bridge/issues/200