Support building with boost asio (#247)
### Public-Facing Changes

Support building with boost asio

### Description
Avoids a hard dependency on standalone asio and provides users with more
flexibility when using foxglove_bridge in a bigger project. Related:
https://github.com/foxglove/ws-protocol/issues/483

Successfully built and tested melodic & noetic with boost asio. I think
it would make sense to add a test for this as well.