Communicate double / double array parameters with type info, explicitly cast when set from integer (#256)
### Public-Facing Changes
Communicate double / double array parameters with type info, explicitly
cast when set from integer

### Description
Implements https://github.com/foxglove/ws-protocol/pull/519

- When sending parameter values to the client, the server will include
type information in case a parameter is of type `PARAMETER_DOUBLE`
(`"type": "float64"`) or an `PARAMETER_ARRAY` of `PARAMETER_DOUBLE`
(`"type": "float64_array"`)
- When receiving a `setParameters` request from the client, the server
checks if the parameter `type` field is present and attempts to cast the
parameter value (or values in case it's an array) to a `double` in case
the type is `float64` / `float64_array`. This is mainly relevant for
ROS2 as setting a double parameter to an integer will raise an
exception. In ROS1, one can change the parameter type on the fly