send service call failure operation (#298)
### Changelog
None

### Docs

None

### Description
Similar to https://github.com/foxglove/ws-protocol/pull/743, this PR
implements the recently added `serviceCallFailure` operation that was
added inhttps://github.com/foxglove/ws-protocol/pull/733. This operation
is sent to a client when the service call failed or if the service could
not be found to avoid that the client potentially waits forever for a
service response message.

---------

Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>