feat: add send buffer queue option (#5)
* Fix "no matching function" error on yocto kirkstone (#331)

### Changelog
None
### Docs
None
### Description

#330 Fixes for Yocto issue

Fix for compilation error on yocto Kirkstone.

> error: no matching function for call to 'max(long unsigned int, size_t)

* bump to 0.8.2 (#332)

Bumping to 0.8.2 for #331

* Add missing functional include in message_definition_cache.cpp (#334)

### Changelog

message_definition_cache.cpp uses std::function, so it should include
the functional STL header

### Docs

None

### Description

message_definition_cache.cpp uses std::function, so it should include
the functional STL header

* add best_effort_qos_topic_whitelist param (#329)

### Changelog
Adds a `qos_reliablity` param to the ros2 bridge that allows the default
reliability to be set via param. Defaults to "automatic" to retain
existing behavior.

### Docs

N/A

### Description

The web-socket bridge runs on TCP which causes issues on flakey internet
connections. There is some discussion here:
https://github.com/orgs/foxglove/discussions/15

This PR adds the ability to specify "qos_reliability" of "best_effort"
which is useful when running a bridge on a different computer to the
robot for example:

**ROBOT COMPUTER** -> DDS/UDP -> **REMOTE COMPUTER** -> Websocket/TCP ->
Foxglove / Browser

---------

Co-authored-by: Hans-Joachim Krauch <achim-k@users.noreply.github.com>

* v0.8.3 (#336)

### Changelog
* add best_effort_qos_topic_whitelist param (#329)
* Add missing functional include in message_definition_cache.cpp (#334)

* feat: add send_buffer_queue param

* chore: lint fixes

---------

Co-authored-by: Graham Harison <gupyfish@gmail.com>
Co-authored-by: Jacob Bandes-Storch <jacob@foxglove.dev>
Co-authored-by: Silvio Traversaro <silvio@traversaro.it>
Co-authored-by: Hans-Joachim Krauch <achim-k@users.noreply.github.com>