Move components from legacy ws-protocol to new namespace (#357)
### Changelog
None

### Docs
None

### Description
This mechanical refactoring PR is in preparation for migrating away from
the legacy foxglove/ws-protocol implementation within the ROS bridge in
favor of the SDK. Including headers from the SDK results in build errors
due to conflicting symbols in the `foxglove` namespace. This PR
introduces a new namespace, `foxglove_ws`, to provide separation, and
migrates the files in `foxglove_bridge_base` that were copied from
foxglvoe/ws-protocol to use this namespace.

Tested that ros1 and ros2 implementations compile successfully.

<!-- In addition to unit tests, describe any manual testing you did to
validate this change. -->

<table><tr><th>Before</th><th>After</th></tr><tr><td>

Code from foxglove/ws-protocol in `foxglove_bridge_base` is in the
`foxglove` namespace.

</td><td>

Code from foxglove/ws-protocol in `foxglove_bridge_base` is in the
`foxglove_ws` namespace.

</td></tr></table>

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->