Set up clangd for vscode dev container (#361)
### Changelog
None

### Docs
None

### Description
Developer quality-of-life change to install `clangd-14` in the VSCode
dev container, configure it to provide `clangd`, and configure CMake to
output `compile_commands.json` when a debug build configuration is
selected.

Enables IntelliSense-like code navigation, in-editor checking, and
semantic completion capabilities in VSCode.

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->