Set up build files and working copy of ros2_foxglove_bridge for SDK (#362)
### Changelog
None

### Docs
None

### Description
Make a copy of `ros2_foxglove_bridge/` in the
`ros2_foxglove_bridge_sdk/` directory, and introduce a build switch
(`USE_FOXGLOVE_SDK`) to determine which will be used as the source
directory for building the `foxglove_bridge_component` target. There are
also changes to the CMakeLists.txt file to enable fetching the C++ SDK
and configuring it as a link dependency if `USE_FOXGLOVE_SDK` is
enabled.

This is intended to create a workspace for developing the SDK version of
the bridge in `main` with incremental and easily-reviewable PRs and
without needing to maintain a long-lived feature integration branch.
Isolating the source files and build flags should allow for this without
affecting production builds.

Tested that the bridge builds as normal if `USE_FOXGLOVE_SDK` is set to
false, and was also able to successfully build in Docker with the SDK
enabled (i.e. `make USE_FOXGLOVE_SDK=ON jazzy`).

<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->

Fixes: FG-12086