Implement server and client message publishing using the SDK server implementation (#364)
### Changelog
None

### Docs
None

### Description
Use the Foxglove SDK to implement ROS message passing from the server to
connected WebSocket clients.

Implement capability to listen to advertisements of channels from
connected clients and pass them through to ROS.

This branch currently depends on changes in foxglove/foxglove-sdk that
are not yet part of a released artifact. Currently have built artifacts
temporarily checked in as LFS objects for development, but that will be
reversed once a release is made.

Note that some of the existing ws-protocol code still exists in
commented-out form for my reference as I develop. I plan on cleaning
this up incrementally as the majority of the code in
`ros_foxglove_bridge.cpp` becomes SDK-ified.

Existing suite of server publication and client advertising unit tests
pass, including cases that test latched topic functionality. Verified
that an existing mcap file from nuScenes can be played and viewed in the
Foxglove webapp using `ros2 bag play`.



https://github.com/user-attachments/assets/75348110-439b-4295-a6d2-cb0056af2124


<!-- If necessary, link relevant Linear or Github issues. Use `Fixes:
foxglove/repo#1234` to auto-close the Github issue or Fixes: FG-### for
Linear isses. -->
Fixes: FG-12233
Fixes: FG-12282